backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend API for amazon scraper"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi[standard]>=0.124.4",
    "lxml>=6.0.2",
    "pydantic-settings>=2.12.0",
    "requests>=2.32.5",
    "sqlmodel>=0.0.27",
]
